backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==3.2.0
    # via
    #   backend
    #   passlib
beautifulsoup4==4.13.5
    # via backend
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   nltk
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==45.0.7
    # via python-jose
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.115.12
    # via backend
fastapi-cli==0.0.10
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
greenlet==3.0.3
    # via playwright
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.9 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
huggingface-hub==0.34.4
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
joblib==1.5.2
    # via
    #   nltk
    #   scikit-learn
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
motor==3.4.0
    # via backend
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
nltk==3.9.1
    # via textstat
numpy==2.3.2
    # via
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.4.5.8 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu12==9.1.0.70 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu12==11.2.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu12==10.3.5.147 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu12==11.6.1.9 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu12==12.3.1.170 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvjitlink-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
passlib==1.7.4
    # via backend
pillow==10.3.0
    # via backend
playwright==1.44.0
    # via backend
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   backend
    #   fastapi
    #   fastapi-cloud-cli
pydantic-settings==2.2.1
    # via backend
pydantic-core==2.18.2
    # via pydantic
pyee==11.1.0
    # via playwright
pygments==2.19.2
    # via rich
pymongo==4.6.3
    # via
    #   backend
    #   motor
pyphen==0.17.2
    # via textstat
python-dotenv==1.0.1
    # via
    #   backend
    #   uvicorn
python-jose==3.3.0
    # via backend
python-multipart==0.0.18
    # via
    #   backend
    #   fastapi
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2025.9.1
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   backend
    #   huggingface-hub
    #   transformers
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
rsa==4.9.1
    # via python-jose
safetensors==0.6.2
    # via transformers
scikit-learn==1.4.2
    # via backend
scipy==1.16.1
    # via scikit-learn
sentry-sdk==2.37.0
    # via fastapi-cloud-cli
setuptools==80.9.0
    # via
    #   textstat
    #   torch
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   bcrypt
    #   ecdsa
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
starlette==0.46.2
    # via fastapi
sympy==1.13.1
    # via torch
textstat==0.7.10
    # via backend
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
torch==2.5.0
    # via backend
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   transformers
transformers==4.41.1
    # via backend
triton==3.1.0 ; python_full_version < '3.13' and platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typer==0.17.4
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   pyee
    #   rich-toolkit
    #   torch
    #   typer
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
